+++ Source/FarmingHysteresis.SmartFarming/Constants.cs
@@ -1,0 +1,6 @@
+namespace FarmingHysteresis.SmartFarming;
+
+internal static class Constants
+{
+    internal const string Id = "FarmingHysteresis.SmartFarming";
+}